Where to begin.
Major Code Cleanup and overhaul.
DatElements now contain a lot more functionality, and DatWrapper has been modified to wrap a standard DatElement as DatElementWrapper.

HexBox now supports keyboard input, editing, color section highlighting.
HexViewWIndow now displays data as int, and as float, and allows one to adjust the number of bytes per line.

Tabuu supports saving, but only in place for now.
This code needs to be cleaned up a bit. It works for now, and Falco400 tested a modified .pac for me.